Fix GetResourceType(int) silent-wrong-stub cluster (3 files)
game/BuildingMgr.cpp and core/BuildingMgrObjectGroup.cpp declared
GetResourceType(int) instead of the real GetResourceType(unsigned int)
(0x446030, Ghidra-verified) - since C++ mangling doesn't encode return
type, this bound clean but silently to shared/defsym_stubs.cpp's
void-returning no-op instead of the real implementation, reading garbage
resource types. ui/UI_ChildWindow.cpp had the same wrong param type AND
had it wrongly declared inside an extern "C" block (dead on host build
via #ifdef _WIN32, but wrong for the mingw typecheck target). Also fixed
a bogus address annotation in world/EditorState.cpp (0x45AAA0 - no
function there - should be 0x446030) and removed the two now-dead stubs
(GetResourceType(int) and a zero-arg GetResourceType(), zero referrers
confirmed via nm). call 0 count unchanged (287 - not a call-0 site).

+++ core/BuildingMgrObjectGroup.cpp
@@ -10,7 +10,7 @@ extern uint32_t g_game_time;                                    // 0x4a99b4
 extern BuildingMgr* g_building_mgr;                             // 0x485448 — host-constructed singleton
 extern void* g_resmgr;                                          // 0x4855e8
 extern void* ResourceManager_GetById(void*, int);                // 0x446ea0
-extern uint8_t GetResourceType(int);                             // 0x446030
+extern unsigned int GetResourceType(unsigned int);                // 0x446030
 extern tm* CRT_localtime(uint32_t*);                             // 0x4674e0
 extern int Game_CheckTimeInRange(int*, int*, int*);              // 0x412710
 extern BOOL Building_CheckPlacement(Building*, int, int);        // 0x433860, vtable [21]
